Added Statement End Date Calculations in calculate.js
The getEndDateFromPeriod function calculates the statement end date from
a user-given period length (i.e.: 30-days of transactions).

The getEndDateFromCycle function uses the startDate to calculate the
cycle date as the end date. The cycle date only falls on business days
and non-holidays. Typically it is near the end of the month on the 28th.
Sometimes Thanksgivings and weekends push the date earlier. A function
to determine whether the end date is conflicting with Thanksgiving
holiday has been made and another function to adjust the end date to
cancel the conflict has also been made.